+++ xfinity-usage/xfinity_usage_addon.py
@@ -24,6 +24,7 @@ from time import sleep
 from paho.mqtt import client as mqtt
 from pathlib import Path
 from playwright.async_api import async_playwright, Playwright, Route, Response, Request, Frame, Page, expect
+from playwright_stealth import StealthConfig, stealth_async
 
 # Browser mode
 HEADLESS = json.loads(os.environ.get('HEADLESS', 'true').lower()) # Convert HEADLESS string into boolean
@@ -439,10 +440,11 @@ class XfinityUsage ():
 
         #logger.info(f"Launching {textwrap.shorten(self.device['user_agent'], width=77, placeholder='...')}")
         
-        self.firefox_user_prefs={'webgl.disabled': True, 'network.http.http2.enabled': False}
+        ##self.firefox_user_prefs={'webgl.disabled': True, 'network.http.http2.enabled': False}
         #self.firefox_user_prefs={'webgl.disabled': True}
         #self.firefox_user_prefs={'webgl.disabled': False}
-        self.webdriver_script = "delete Object.getPrototypeOf(navigator).webdriver"
+        #self.firefox_user_prefs={}
+        ##self.webdriver_script = "delete Object.getPrototypeOf(navigator).webdriver"
         #self.webdriver_script = ""
 
         #self.browser = playwright.firefox.launch(headless=False,slow_mo=1000,firefox_user_prefs=self.firefox_user_prefs)
@@ -457,8 +459,10 @@ class XfinityUsage ():
 
         #self.context = playwright.firefox.launch_persistent_context(profile_path,headless=False,firefox_user_prefs=self.firefox_user_prefs,**self.device)
         #self.context = playwright.firefox.launch_persistent_context(profile_path,headless=False,firefox_user_prefs=self.firefox_user_prefs,**self.device)
-        self.context = await self.playwright.firefox.launch_persistent_context(self.profile_path,headless=HEADLESS,firefox_user_prefs=self.firefox_user_prefs,**self.device)
+        ##self.context = await self.playwright.firefox.launch_persistent_context(self.profile_path,headless=HEADLESS,firefox_user_prefs=self.firefox_user_prefs,**self.device)
 
+        self.browser = await self.playwright.firefox.launch(headless=HEADLESS)
+        self.context = await self.browser.new_context()
 
         # Block unnecessary requests
         await self.context.route("**/*", lambda route: self.abort_route(route))
@@ -470,16 +474,40 @@ class XfinityUsage ():
         self.context.on("requestfailed", self.check_requestfailed)
         self.context.on("requestfinished", self.check_requestfinished)
 
-
+        await stealth_async(
+            self.context,
+            StealthConfig(
+                webdriver=True,
+                webgl_vendor=True,
+                chrome_app=False,
+                chrome_csi=False,
+                chrome_load_times=False,
+                chrome_runtime=False,
+                iframe_content_window=True,
+                media_codecs=True,
+                navigator_hardware_concurrency=4,
+                navigator_languages=False,
+                navigator_permissions=True,
+                navigator_platform=False,
+                navigator_plugins=True,
+                navigator_user_agent=False,
+                navigator_vendor=False,
+                outerdimensions=True,
+                hairline=False,
+            ),
+        )
+        
         #self.page = await self.context.new_page()
         self.page = await self.get_new_page()
-
+        
         logger.info(f"Launching {textwrap.shorten(await self.page.evaluate('navigator.userAgent'), width=77, placeholder='...')}")
 
         if  DEBUG_SUPPORT and \
             os.path.exists('/config/'):
             self.page.on("console", lambda consolemessage: debug_support_logger.debug(f"Console Message: {consolemessage.text}"))
             self.page.on("pageerror", self.check_pageerror)
+            await self.page.goto("https://bot.sannysoft.com/", wait_until="networkidle")
+            await self.page.screenshot(path=f"/config/{datetime.now().strftime('%Y-%m-%dT%H:%M:%S.%f')}-sannysoft.png", full_page=True)
         self.page.on("close", self.check_close)
         self.page.on("domcontentloaded", self.check_domcontentloaded)
         self.page.on("frameattached", self.check_frameattached)
@@ -504,7 +532,7 @@ class XfinityUsage ():
         expect.set_options(timeout=self.timeout)
 
         # Help reduce bot detection
-        await _page.add_init_script(self.webdriver_script)
+        ##await _page.add_init_script(self.webdriver_script)
 
         return _page
 
@@ -566,7 +594,7 @@ class XfinityUsage ():
 
     async def abort_route(self, route: Route) :
         # Necessary Xfinity domains
-        good_xfinity_domains = ['*.xfinity.com', '*.comcast.net', 'static.cimcontent.net', '*.codebig2.net']
+        good_xfinity_domains = ['*.xfinity.com', '*.comcast.net', 'static.cimcontent.net', '*.codebig2.net', 'bot.sannysoft.com']
         regex_good_xfinity_domains = ['xfinity.com', 'comcast.net', 'static.cimcontent.net', 'codebig2.net']
 
         #good_xfinity_domains = ['*.xfinity.com', '*.comcast.net', 'static.cimcontent.net', '*.codebig2.net', '*']
@@ -1376,7 +1404,8 @@ async def run_playwright() -> None:
             
             if is_mqtt_available() and mqtt_client.is_connected_mqtt():
                 mqtt_client.publish_mqtt(usage.usage_data)
-
+        except Exception as e:
+            logger.warning(f"Error: {e}")    
         finally:
             await usage.done()