+++ xfinity-usage/CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## 0.1.2
 
+- Uses playwright-stealth to mask headless browser signature. [(csobrinoho:fix-56)](https://github.com/csobrinho/hassio-xfinity-usage/tree/fix-56)
+
 - Dependency Update
     - playwright v1.49.1
+    - playwright-stealth v1.0.6
 
 ## 0.1.1